+++ api/company_test.go
@@ -2,6 +2,7 @@ package api
 
 import (
 	"context"
+	"fmt"
 	"net/http"
 	"os"
 	"testing"
@@ -17,7 +18,7 @@ func TestCompany_Enrich(t *testing.T) {
 
 	// test
 	params := model.EnrichCompanyParams{
-		CompanyParams: model.CompanyParams{Name: "Google, Inc."},
+		CompanyParams: model.CompanyParams{Name: "Google"},
 	}
 	resp, err := company.Enrich(context.Background(), params)
 
@@ -25,6 +26,37 @@ func TestCompany_Enrich(t *testing.T) {
 	assert.NoError(t, err)
 	assert.Equal(t, http.StatusOK, resp.Status)
 	assert.Equal(t, 1998, resp.Founded)
+}
+
+func TestCompany_BulkEnrich(t *testing.T) {
+	// setup
+	company := Company{Client: NewClient(os.Getenv("PDL_API_KEY"), "1.0.0")}
+
+	// test
+	params := model.BulkEnrichCompanyParams{
+		BaseParams: model.BaseParams{Pretty: true},
+		Requests: []model.BulkEnrichSingleCompanyParams{
+			{
+				Params: model.CompanyParams{
+					Profile: "linkedin.com/company/peopledatalabs/",
+				},
+			},
+			{
+				Params: model.CompanyParams{
+					Profile: "https://www.linkedin.com/company/apple/",
+				},
+			},
+		},
+	}
+	resp, err := company.BulkEnrich(context.Background(), params)
+
+	fmt.Println(resp)
+	// assertions
+	assert.NoError(t, err)
+	assert.Equal(t, http.StatusOK, resp[0].Status)
+	assert.Equal(t, http.StatusOK, resp[1].Status)
+	assert.Equal(t, "People Data Labs", resp[0].DisplayName)
+	assert.Equal(t, "Apple", resp[1].DisplayName)
 }
 
 func TestCompany_Clean(t *testing.T) {